requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
numpy>=1.19.5
pandas>=1.3.0
tqdm>=4.62.0
faiss-cpu>=1.7.3
python-multipart>=0.0.5
sentence-transformers>=2.2.2
python-dotenv>=1.0.0